Implement clientside caching of server resources
Reduces the number of requests we have to make to the backend. Improving performance on slower networks.

As a result of the changes here:
Need to redevelop the method the front end uses to get terminal output from the back end. Best way of doing this will be to use sockets.
When a server is deleted the front end needs to prevent rendering of a null server.